react-app: consolidate & adopt proposal utils refs oursky/likedao#101
