+++ react-app/src/i18n/translations/zh.json
@@ -36,6 +36,12 @@
   "OverviewScreen.requestState.error": "Failed to fetch data, please try again later.",
   "PaginationController.next": "Next",
   "PaginationController.previous": "Previous",
+  "Proposal.status.depositPeriod": "Deposit Period",
+  "Proposal.status.failed": "Failed",
+  "Proposal.status.invalid": "Invalid",
+  "Proposal.status.passed": "Passed",
+  "Proposal.status.rejected": "Rejected",
+  "Proposal.status.votingPeriod": "Voting Period",
   "ProposalDetail.addReaction": "Add Reaction",
   "ProposalDetail.copyLink": "Copy Link",
   "ProposalDetail.deposit": "Deposit",
@@ -58,12 +64,6 @@
   "ProposalScreen.newProposal": "New proposal",
   "ProposalScreen.noProposals": "No Proposals",
   "ProposalScreen.proposalCard.viewDetails": "View details",
-  "ProposalScreen.proposalStatus.depositPeriod": "Deposit Period",
-  "ProposalScreen.proposalStatus.failed": "Failed",
-  "ProposalScreen.proposalStatus.invalid": "Invalid",
-  "ProposalScreen.proposalStatus.passed": "Passed",
-  "ProposalScreen.proposalStatus.rejected": "Rejected",
-  "ProposalScreen.proposalStatus.votingPeriod": "Voting Period",
   "ProposalScreen.proposalType.cancelSoftwareUpgrade": "Cancel Software Upgrade",
   "ProposalScreen.proposalType.communityPoolSpend": "Community Pool Spend",
   "ProposalScreen.proposalType.parameterChange": "Parameter Change",
